[MIN-107] Refactoring BalanceAnnotation.
