+++ src/App.jsx
@@ -4,10 +4,7 @@ import 'semantic-ui-css/semantic.min.css';
 
 import { loadAccounts, setAccount, checkIsAdmin } from './actions/accounts';
 import { initializeAPI } from './actions/api';
-import {
-	getBalanceAnnotation,
-	resetBalanceAnnotation,
-} from './actions/balanceAnnotation';
+import { getBalanceAnnotation } from './actions/dashboardData';
 import { API_STATE_READY, KEYRING_STATE_READY } from './util/constants';
 import { Tab } from 'semantic-ui-react';
 
@@ -33,7 +30,6 @@ function App(props) {
 
 		getBalanceAnnotation,
 		balanceAnnotation,
-		resetBalanceAnnotation,
 	} = props;
 	const [isInitialized, setIsInitialized] = useState(false);
 
@@ -51,8 +47,6 @@ function App(props) {
 	useEffect(() => {
 		if (currentAccount) {
 			getBalanceAnnotation(currentAccount);
-		} else {
-			resetBalanceAnnotation();
 		}
 	}, [currentAccount]);
 
@@ -120,7 +114,7 @@ const mapStateToProps = (state) => ({
 	keyring: state.account.keyring,
 	isAdmin: state.account.isAdmin,
 	isAdminRequestRunning: state.account.isAdminRequestRunning,
-	balanceAnnotation: state.balanceAnnotation.balanceAnnotation,
+	balanceAnnotation: state.dashboardData.balanceAnnotation,
 });
 const mapDispatchToProps = {
 	loadAccounts,
@@ -128,7 +122,6 @@ const mapDispatchToProps = {
 	setAccount,
 	checkIsAdmin,
 	getBalanceAnnotation,
-	resetBalanceAnnotation,
 };
 
 export default connect(mapStateToProps, mapDispatchToProps)(App);
